memory manager: allocate memory on lower 2GB
Use lower 2GB to allocate the memory when the bitcode is 32-bit.
The allocation on lower 2GB is achieved by memmap. It has the disadventage
that we always allocate whole page, but still better try with a big
memory consumption than crashing.

Without this the KLEE is hitting an assert when we try to bitcast
the 64-bit address to 32-bit value

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>